[ADD] Added method of async write file
